+++ lib/presentation/history/view_models/calendar_view_model.dart
@@ -1,43 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:intl/intl.dart';
+import 'package:mongbi_app/presentation/history/models/calendar_model.dart';
 
-class Calendar {
-  Calendar({DateTime? focusedDay, this.selectedDay, String? dropDownvalue}) {
-    this.focusedDay = focusedDay ?? DateTime.now();
-    this.dropDownvalue =
-        dropDownvalue ?? DateFormat('yyyy년 MM월').format(DateTime.now());
-  }
-
-  late DateTime focusedDay;
-  DateTime? selectedDay;
-  late String dropDownvalue;
-
-  Calendar copyWith({
-    DateTime? focusedDay,
-    DateTime? selectedDay,
-    String? dropDownvalue,
-  }) {
-    return Calendar(
-      focusedDay: focusedDay ?? this.focusedDay,
-      selectedDay: selectedDay ?? this.selectedDay,
-      dropDownvalue: dropDownvalue ?? this.dropDownvalue,
-    );
-  }
-
+class CalendarViewModel extends AutoDisposeNotifier<CalendarModel> {
   @override
-  String toString() {
-    return '''{
-      focusedDay: $focusedDay, 
-      selectedDay: $selectedDay,
-      dropDownvalue: $dropDownvalue
-    }''';
-  }
-}
-
-class CalendarViewModel extends AutoDisposeNotifier<Calendar> {
-  @override
-  Calendar build() {
-    return Calendar();
+  CalendarModel build() {
+    return CalendarModel();
   }
 
   void onChangedCalendar(DateTime date) {
@@ -46,8 +13,6 @@ class CalendarViewModel extends AutoDisposeNotifier<Calendar> {
 
   void onChangeDropDownValue(dynamic value) {
     state = state.copyWith(dropDownvalue: value);
-    print('adfadf');
-    print(state.toString());
   }
 
   void onDaySelected(DateTime selectedDay, DateTime focusedDay) {
@@ -60,6 +25,6 @@ class CalendarViewModel extends AutoDisposeNotifier<Calendar> {
 }
 
 final calendarViewModelProvider =
-    NotifierProvider.autoDispose<CalendarViewModel, Calendar>(() {
+    NotifierProvider.autoDispose<CalendarViewModel, CalendarModel>(() {
       return CalendarViewModel();
     });
